% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculations.R
\name{calculate_final_settlement}
\alias{calculate_final_settlement}
\title{Calculate Final Settlement}
\usage{
calculate_final_settlement(
  all_expense_types,
  people,
  type_totals,
  presence_ratios,
  expense_summary,
  shared_expense_types,
  exceptions
)
}
\arguments{
\item{all_expense_types}{All expense types}

\item{people}{List of people}

\item{type_totals}{Type totals data frame}

\item{presence_ratios}{Presence ratios data frame}

\item{expense_summary}{Expense summary data frame}

\item{shared_expense_types}{Shared expense types}

\item{exceptions}{Exceptions data frame}
}
\value{
Final calculations data frame
}
\description{
Calculate Final Settlement
}
