% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{check_absences_input}
\alias{check_absences_input}
\title{Check Absences Input File Columns and Types}
\usage{
check_absences_input(df)
}
\arguments{
\item{df}{The absences data frame}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Check Absences Input File Columns and Types
}
