% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{check_exceptions_input}
\alias{check_exceptions_input}
\title{Check Exceptions Input File Columns and Types}
\usage{
check_exceptions_input(df)
}
\arguments{
\item{df}{The exceptions data frame}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Check Exceptions Input File Columns and Types
}
