% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{check_expenses_input}
\alias{check_expenses_input}
\title{Check Expenses Input File Columns and Types}
\usage{
check_expenses_input(df)
}
\arguments{
\item{df}{The expenses data frame}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Check Expenses Input File Columns and Types
}
