% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_absences_section}
\alias{create_absences_section}
\title{Create Absences Section for RMD}
\usage{
create_absences_section(absences_summary)
}
\arguments{
\item{absences_summary}{Absences summary data frame}
}
\value{
Character string with absences section
}
\description{
Create Absences Section for RMD
}
