% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_detailed_calculations_section}
\alias{create_detailed_calculations_section}
\title{Create Detailed Calculations Section for RMD}
\usage{
create_detailed_calculations_section(detailed_calculations)
}
\arguments{
\item{detailed_calculations}{Detailed calculations data frame}
}
\value{
Character string with detailed calculations section
}
\description{
Create Detailed Calculations Section for RMD
}
