% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_exceptions_section}
\alias{create_exceptions_section}
\title{Create Exceptions Section for RMD}
\usage{
create_exceptions_section(exceptions_summary)
}
\arguments{
\item{exceptions_summary}{Exceptions summary data frame}
}
\value{
Character string with exceptions section
}
\description{
Create Exceptions Section for RMD
}
