% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_expense_details_section}
\alias{create_expense_details_section}
\title{Create Expense Details Section for RMD}
\usage{
create_expense_details_section(expenses)
}
\arguments{
\item{expenses}{Expenses data frame}
}
\value{
Character string with expense details section
}
\description{
Create Expense Details Section for RMD
}
