% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_final_settlement_section}
\alias{create_final_settlement_section}
\title{Create Final Settlement Section for RMD}
\usage{
create_final_settlement_section(final_settlement)
}
\arguments{
\item{final_settlement}{Final settlement data frame}
}
\value{
Character string with final settlement section
}
\description{
Create Final Settlement Section for RMD
}
