% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{create_summary_by_type_section}
\alias{create_summary_by_type_section}
\title{Create Summary by Type Section for RMD}
\usage{
create_summary_by_type_section(summary_by_type)
}
\arguments{
\item{summary_by_type}{Summary by type data frame}
}
\value{
Character string with summary by type section
}
\description{
Create Summary by Type Section for RMD
}
