% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pdf_report.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Generate PDF Report}
\usage{
generate_pdf_report(
  calculations,
  start_date,
  end_date,
  absences_data,
  exceptions_data,
  file
)
}
\arguments{
\item{calculations}{Calculations result}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{absences_data}{Absences data}

\item{exceptions_data}{Exceptions data}

\item{file}{Output file path}
}
\description{
Generate PDF Report
}
