% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_default_expense_types}
\alias{get_default_expense_types}
\title{Get Default Expense Types
'scale_fill_viridis_d' is not an exported object from 'namespace:viridis'}
\usage{
get_default_expense_types()
}
\value{
Character string with default expense types
}
\description{
Get Default Expense Types
'scale_fill_viridis_d' is not an exported object from 'namespace:viridis'
}
