% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_default_people_list}
\alias{get_default_people_list}
\title{Get Default People List}
\usage{
get_default_people_list()
}
\value{
Character string with default people list
}
\description{
Get Default People List
}
