% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_default_shared_expense_types}
\alias{get_default_shared_expense_types}
\title{Get Default Shared Expense Types}
\usage{
get_default_shared_expense_types()
}
\value{
Character string with default shared expense types
}
\description{
Get Default Shared Expense Types
}
