% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{load_default_from_assets}
\alias{load_default_from_assets}
\title{Load Default Values from Assets Folder}
\usage{
load_default_from_assets(filename, fallback_default)
}
\arguments{
\item{filename}{The filename to load from assets folder}

\item{fallback_default}{Default value to use if file doesn't exist}
}
\value{
Character string with file content or fallback default
}
\description{
Load Default Values from Assets Folder
}
