% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{mod_data_upload_server}
\alias{mod_data_upload_server}
\title{Data Upload Module Server}
\usage{
mod_data_upload_server(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
List of reactive values for uploaded data
}
\description{
Data Upload Module Server
}
