% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{mod_data_upload_ui}
\alias{mod_data_upload_ui}
\title{Data Upload Module UI}
\usage{
mod_data_upload_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
HTML div with data upload inputs
}
\description{
Data Upload Module UI
}
