% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_instructions.R
\name{mod_instructions_server}
\alias{mod_instructions_server}
\title{Instructions Module Server}
\usage{
mod_instructions_server(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Instructions Module Server
}
