% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_instructions.R
\name{mod_instructions_ui}
\alias{mod_instructions_ui}
\title{Instructions Module UI}
\usage{
mod_instructions_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
HTML div with instructions content
}
\description{
Instructions Module UI
}
