% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_long_term.R
\name{mod_long_term_server}
\alias{mod_long_term_server}
\title{Long Term Analysis Module Server}
\usage{
mod_long_term_server(id, expenses_data)
}
\arguments{
\item{id}{Module namespace}

\item{expenses_data}{Reactive expenses data}
}
\description{
Long Term Analysis Module Server
}
