% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_long_term.R
\name{mod_long_term_ui}
\alias{mod_long_term_ui}
\title{Long Term Analysis Module UI}
\usage{
mod_long_term_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
HTML div with long term analysis content
}
\description{
Long Term Analysis Module UI
}
