% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results_server}
\alias{mod_results_server}
\title{Results Module Server}
\usage{
mod_results_server(
  id,
  calculations,
  start_date,
  end_date,
  absences_data,
  exceptions_data
)
}
\arguments{
\item{id}{Module namespace}

\item{calculations}{Reactive calculations result}

\item{start_date}{Reactive start date}

\item{end_date}{Reactive end date}

\item{absences_data}{Reactive absences data}

\item{exceptions_data}{Reactive exceptions data}
}
\description{
Results Module Server
}
