% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results_ui}
\alias{mod_results_ui}
\title{Results Module UI}
\usage{
mod_results_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
HTML div with results content
}
\description{
Results Module UI
}
