% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_setup.R
\name{mod_setup_server}
\alias{mod_setup_server}
\title{Setup Module Server}
\usage{
mod_setup_server(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
Reactive values with calculation results and settings
}
\description{
Setup Module Server
}
