% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculations.R
\name{process_absences}
\alias{process_absences}
\title{Process Absences Data}
\usage{
process_absences(absences_data, people)
}
\arguments{
\item{absences_data}{Raw absences data}

\item{people}{List of people}
}
\value{
Processed absences data frame
}
\description{
Process Absences Data
}
