% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculations.R
\name{process_exceptions}
\alias{process_exceptions}
\title{Process Exceptions Data}
\usage{
process_exceptions(exceptions_data)
}
\arguments{
\item{exceptions_data}{Raw exceptions data}
}
\value{
Processed exceptions data frame
}
\description{
Process Exceptions Data
}
