% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculations.R
\name{validate_data}
\alias{validate_data}
\title{Validate Input Data}
\usage{
validate_data(
  expenses,
  people,
  all_expense_types,
  absences_data,
  exceptions_data,
  total_days
)
}
\arguments{
\item{expenses}{Expenses data frame}

\item{people}{List of people}

\item{all_expense_types}{List of all expense types}

\item{absences_data}{Absences data frame}

\item{exceptions_data}{Exceptions data frame}

\item{total_days}{Total days in period}
}
\value{
List with errors and warnings
}
\description{
Validate Input Data
}
